update with plugin composer
